doc/developers-app: Add working way to link archives
Linking an archive with OBJS can fail a build as archives can be listed before
the object files that actually need them. By using ALIB variable the archives
are placed in a group that the linker will search multiple times for unresolved
symbols. This is why the documentation for OBJS variable is changed to ALIB.

Signed-off-by: Rochet2 <tqbattlenet@gmail.com>